fix discount discount price, fix home filter by news, sales
