+++ app/src/main/java/com/goldenowl/ecommerce/adapter/CategoryProductListAdapter.kt
@@ -104,7 +104,6 @@ class CategoryProductListAdapter(private val mLayoutManager: GridLayoutManager, 
     }
 
     override fun onBindViewHolder(holder: CategoryProductViewHolder, position: Int) {
-        Log.d(TAG, "onBindViewHolder: binding $position")
         val product = mListProductData[position].product
         val favorite = mListProductData[position].favorite
 
@@ -141,7 +140,6 @@ class CategoryProductListAdapter(private val mLayoutManager: GridLayoutManager, 
                 val text = product.getOriginPrice().toString() + "$"
                 holder.originPrice?.strike(text)
             } else {
-                Log.d(TAG, "onBindViewHolder: NO SALE")
                 holder.tvDiscountPercent?.visibility = View.INVISIBLE
                 holder.discountPrice?.visibility = View.INVISIBLE
                 holder.originPrice?.text = product.getOriginPrice().toString() + "$"
