+++ app/src/main/java/com/goldenowl/ecommerce/ui/global/home/HomeFragment.kt
@@ -14,6 +14,8 @@ import com.goldenowl.ecommerce.models.data.Product
 import com.goldenowl.ecommerce.models.data.ProductData
 import com.goldenowl.ecommerce.ui.global.BaseHomeFragment
 import com.goldenowl.ecommerce.ui.global.MainActivity
+import com.goldenowl.ecommerce.ui.global.bottomsheet.BottomSheetInsertFavorite
+import com.goldenowl.ecommerce.ui.global.favorites.FavoritesFragment
 import com.goldenowl.ecommerce.utils.BaseLoadingStatus
 import com.goldenowl.ecommerce.utils.Utils.getColor
 
@@ -56,7 +58,14 @@ class HomeFragment : BaseHomeFragment<FragmentHomeBinding>() {
         Log.d(TAG, "setViews: started")
         val listener = object : HomeProductListAdapter.IClickListener {
             override fun onClickFavorite(product: Product, favorite: Favorite?) {
-                Log.d(TAG, "onClickFavorite: insert favorite")
+                Log.d(CategoryFragment.TAG, "onClickFavorite: $favorite")
+                if (favorite == null) {
+                    Log.d(FavoritesFragment.TAG, "onClickFavorite: insert favorite")
+                    toggleBottomSheetAddToFavorite(product)
+                } else {
+                    Log.d(FavoritesFragment.TAG, "onClickFavorite: remove favorite")
+                    viewModel.removeFavorite(favorite!!)
+                }
             }
         }
         salesListAdapter = HomeProductListAdapter(listener)
@@ -95,5 +104,9 @@ class HomeFragment : BaseHomeFragment<FragmentHomeBinding>() {
             setExpandedTitleColor(getColor(context, R.color.white) ?: 0xFFFFFF)
         }
     }
-
+    private fun toggleBottomSheetAddToFavorite(product: Product) {
+        val modalBottomSheet = BottomSheetInsertFavorite(product, viewModel)
+        modalBottomSheet.enterTransition = View.GONE
+        modalBottomSheet.show(parentFragmentManager, BottomSheetInsertFavorite.TAG)
+    }
 }
