+++ app/src/main/res/layout/item_product_list_category.xml
@@ -102,15 +102,34 @@
 
                 </LinearLayout>
 
-                <TextView
-                    android:id="@+id/product_origin_price"
-                    android:layout_width="wrap_content"
+                <LinearLayout
+                    android:layout_width="match_parent"
                     android:layout_height="wrap_content"
-                    android:layout_marginTop="9dp"
-                    android:text="15$"
-                    android:textColor="@color/text_title_color"
-                    android:textSize="14sp"
-                    android:textStyle="bold" />
+                        android:layout_marginTop="9dp"
+                    android:orientation="horizontal">
+
+                    <TextView
+                        android:id="@+id/product_origin_price"
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:text="15$"
+                        android:textColor="@color/text_title_color"
+                        android:textSize="14sp"
+                        android:textStyle="bold" />
+
+                    <TextView
+                        android:id="@+id/product_discount_price"
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:layout_marginStart="3dp"
+                        android:text="15$"
+                        android:textColor="@color/red_dark"
+                        android:textSize="14sp"
+                        android:textStyle="bold"
+                        app:layout_constraintBottom_toBottomOf="parent"
+                        app:layout_constraintStart_toEndOf="@+id/product_origin_price"
+                        app:layout_constraintTop_toTopOf="parent" />
+                </LinearLayout>
             </LinearLayout>
