+++ app/src/main/res/layout/item_product_list_category_grid.xml
@@ -119,7 +119,7 @@
         app:layout_constraintTop_toBottomOf="@+id/product_name">
 
         <TextView
-            android:id="@+id/product_old_price"
+            android:id="@+id/product_origin_price"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="15$"
@@ -127,7 +127,7 @@
             android:textSize="14sp" />
 
         <TextView
-            android:id="@+id/product_origin_price"
+            android:id="@+id/product_discount_price"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginStart="4dp"
